+++ Intent.Modules.Java.Services.CRUD/Decorators/ImplementationStrategies/GetAllImplementationStrategy.cs
@@ -1,11 +1,8 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using Intent.Modelers.Domain.Api;
-using Intent.Modelers.Services.Api;
 using Intent.Modules.Common.Templates;
-using Intent.Modules.Java.Services.Templates;
 using Intent.Modules.Java.Services.Templates.ServiceImplementation;
-using Intent.Modules.Java.Spring.Data.Repositories.Templates.EntityRepository;
 using OperationModel = Intent.Modelers.Services.Api.OperationModel;
 
 namespace Intent.Modules.Java.Services.CRUD.Decorators.ImplementationStrategies
@@ -26,7 +23,7 @@ namespace Intent.Modules.Java.Services.CRUD.Decorators.ImplementationStrategies
                 return false;
             }
 
-            if (!(operationModel?.TypeReference?.IsCollection ?? false))
+            if (operationModel.TypeReference?.IsCollection != true)
             {
                 return false;
             }
@@ -52,19 +49,20 @@ namespace Intent.Modules.Java.Services.CRUD.Decorators.ImplementationStrategies
 
         public string GetImplementation(ClassModel domainModel, OperationModel operationModel)
         {
-            var dto = operationModel.TypeReference.Element.AsDTOModel();
-            return $@"var {domainModel.Name.ToCamelCase().ToPluralName()} = {domainModel.Name.ToCamelCase()}Repository.findAll();
-        return {_decorator.Template.GetDataTransferModelName(dto)}.mapFrom{domainModel.Name.ToPluralName()}({domainModel.Name.ToCamelCase().ToPluralName()}, mapper);";
+            var domainType = _decorator.GetDomainTypeName(domainModel);
+            var domainTypeCamelCased = domainType.ToCamelCase();
+            var domainTypePascalCased = domainType.ToPascalCase();
+            var repositoryFieldName = _decorator.GetRepositoryDependency(domainModel).Name;
+            var dtoType = _decorator.GetDtoTypeName(operationModel.TypeReference.Element);
+
+            return $@"var {domainTypeCamelCased.Pluralize()} = {repositoryFieldName}.findAll();
+        return {dtoType}.mapFrom{domainTypePascalCased.Pluralize()}({domainTypeCamelCased.Pluralize()}, mapper);";
         }
 
         public IEnumerable<ClassDependency> GetRequiredServices(ClassModel targetEntity)
         {
-            var repo = _decorator.Template.GetTypeName(EntityRepositoryTemplate.TemplateId, targetEntity);
-            return new[]
-            {
-                new ClassDependency(repo, repo.ToCamelCase()),
-                new ClassDependency("org.modelmapper.ModelMapper", "mapper"), 
-            };
-        }   
+            yield return _decorator.GetRepositoryDependency(targetEntity);
+            yield return new ClassDependency("org.modelmapper.ModelMapper", "mapper");
+        }
     }
 }